Rename journalctl to journald
