--- plugin/input/journalctl/init.go
@@ -1,3 +1,0 @@
-package journalctl
-
-// this file used to import journalctl plugin without errors on the darwin platform
