+++ plugin/input/journald/README.md
@@ -1,5 +1,5 @@
 # Journal.d plugin
-Reads `journalctl` output.
+Reads `journald` output.
 
 ### Config params
 **`offsets_file`** *`string`* *`required`* 
@@ -10,10 +10,10 @@ The filename to store offsets of processed messages.
 
 **`journal_args`** *`[]string`* *`default=-f -a`* 
 
-Additional args for `journalctl`.
+Additional args for `journald`.
 Plugin forces "-o json" and "-c *cursor*" or "-n all", otherwise
 you can use any additional args.
-> Have a look at https://man7.org/linux/man-pages/man1/journalctl.1.html
+> Have a look at https://man7.org/linux/man-pages/man1/journald.1.html
 
 <br>
