+++ plugin/input/journald/init.go
@@ -1,0 +1,3 @@
+package journald
+
+// this file used to import journald plugin without errors on the darwin platform
