+++ plugin/input/journald/journald.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package journalctl
+package journald
 
 import (
 	"strings"
@@ -15,7 +15,7 @@ import (
 )
 
 /*{ introduction
-Reads `journalctl` output.
+Reads `journald` output.
 }*/
 
 type Plugin struct {
@@ -28,9 +28,9 @@ type Plugin struct {
 
 	//  plugin metrics
 
-	offsetErrorsMetric        *prometheus.CounterVec
-	journalCtlStopErrorMetric *prometheus.CounterVec
-	readerErrorsMetric        *prometheus.CounterVec
+	offsetErrorsMetric      *prometheus.CounterVec
+	journaldStopErrorMetric *prometheus.CounterVec
+	readerErrorsMetric      *prometheus.CounterVec
 }
 
 type Config struct {
@@ -44,10 +44,10 @@ type Config struct {
 
 	// > @3@4@5@6
 	// >
-	// > Additional args for `journalctl`.
+	// > Additional args for `journald`.
 	// > Plugin forces "-o json" and "-c *cursor*" or "-n all", otherwise
 	// > you can use any additional args.
-	// >> Have a look at https://man7.org/linux/man-pages/man1/journalctl.1.html
+	// >> Have a look at https://man7.org/linux/man-pages/man1/journald.1.html
 	JournalArgs []string `json:"journal_args" default:"-f -a"` // *
 
 	// for testing mostly
@@ -65,14 +65,14 @@ func (o *offsetInfo) set(cursor string) {
 }
 
 func (p *Plugin) Write(bytes []byte) (int, error) {
-	p.params.Controller.In(0, "journalctl", p.currentOffset, bytes, false)
+	p.params.Controller.In(0, "journald", p.currentOffset, bytes, false)
 	p.currentOffset++
 	return len(bytes), nil
 }
 
 func init() {
 	fd.DefaultPluginRegistry.RegisterInput(&pipeline.PluginStaticInfo{
-		Type:    "journalctl",
+		Type:    "journald",
 		Factory: Factory,
 	})
 }
@@ -108,16 +108,16 @@ func (p *Plugin) Start(config pipeline.AnyConfig, params *pipeline.InputPluginPa
 }
 
 func (p *Plugin) registerMetrics(ctl *metric.Ctl) {
-	p.offsetErrorsMetric = ctl.RegisterCounter("input_journalctl_offset_errors", "Number of errors occurred when saving/loading offset")
-	p.journalCtlStopErrorMetric = ctl.RegisterCounter("input_journalctl_stop_errors", "Total journalctl stop errors")
-	p.readerErrorsMetric = ctl.RegisterCounter("input_journalctl_reader_errors", "Total reader errors")
+	p.offsetErrorsMetric = ctl.RegisterCounter("input_journald_offset_errors", "Number of errors occurred when saving/loading offset")
+	p.journaldStopErrorMetric = ctl.RegisterCounter("input_journald_stop_errors", "Total journald stop errors")
+	p.readerErrorsMetric = ctl.RegisterCounter("input_journald_reader_errors", "Total reader errors")
 }
 
 func (p *Plugin) Stop() {
 	err := p.reader.stop()
 	if err != nil {
-		p.journalCtlStopErrorMetric.WithLabelValues().Inc()
-		p.logger.Error("can't stop journalctl cmd", zap.Error(err))
+		p.journaldStopErrorMetric.WithLabelValues().Inc()
+		p.logger.Error("can't stop journald cmd", zap.Error(err))
 	}
 
 	offsets := *p.offInfo.Load()
