+++ plugin/input/journald/journald_test.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package journalctl
+package journald
 
 import (
 	"path/filepath"
