+++ plugin/input/journald/reader.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package journalctl
+package journald
 
 import (
 	"bufio"
